spi: fix deadlock
On removal, some devices would try to issue commands. This didn't work
with the driver because if the ch341 was physically removed, it would
take the spi_lock to both do the removal, and issue commands.

Fix by using a mutex for transfers, and another for device accounting.

This also fixes races with device registration / deregistration.