Replace transfer_one() with transfer_one_message()
For speed, the driver needs to send data in large chunks, which can't
happen when the caller is splitting its request in tiny bits. So
provide transfer_one_message() where the driver can reassemble the
transfer requests.

This speeds up reading from a W25Q flash from 27KB/s to 104KB/s when
using the in-kernel spi-nor driver.